Removes farm context and updates gamma context
