+++ src/context/index.ts
@@ -12,5 +12,4 @@ export * from './wallet'
 export * from './reward_toggle'
 export * from './price_feed_farm'
 export * from './ssl'
-export * from './farm'
 export * from './gamma'
