+++ src/pages/FarmV4/DepositWithdrawAccordion.tsx
@@ -1,10 +1,10 @@
 import { FC, ReactElement } from 'react'
-import { useFarmContext } from '@/context'
+import {  useGamma } from '@/context'
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from 'gfx-component-lib'
 import { PoolStats } from './PoolStats'
 
 const DepositWithdrawAccordion: FC = (): ReactElement => {
-  const { selectedCard } = useFarmContext()
+  const { selectedCard } = useGamma()
     return(
         <Accordion collapsible type={'multiple'} variant='default'
                    className="dark:bg-black-1 bg-grey-5 mx-2.5 my-3 !rounded-[4px]">
