+++ src/pages/FarmV4/DepositWithdrawHeader.tsx
@@ -1,10 +1,10 @@
-import { useFarmContext } from '@/context'
+import { useGamma } from '@/context'
 import { FC } from 'react'
 import { Button, DialogTitle, Icon } from 'gfx-component-lib'
 import RewardsClose from '@/assets/rewards_close.svg?react'
 
 export const DepositWithdrawHeader: FC = (): JSX.Element => {
-  const { selectedCard, setOperationPending } = useFarmContext()
+  const { selectedCard, setOperationPending } = useGamma()
   return (
     <DialogTitle
       className="w-full h-14 flex flex-row items-center border-b
