+++ src/pages/FarmV4/ReviewConfirm.tsx
@@ -1,20 +1,20 @@
-/* eslint-disable */
-import { useFarmContext } from '@/context'
+import { useGamma } from '@/context'
 import { FC, ReactElement } from 'react'
 import { Container, Icon } from 'gfx-component-lib'
 import DepositWithdrawLabel from './DepositWithdrawLabel'
 
 export const ReviewConfirm: FC = (): ReactElement => {
-    const { selectedCard } = useFarmContext()
+    const { selectedCard } = useGamma()
     return (
         <>
             <DepositWithdrawLabel text="2. Review and Confirm" />
-            {/*<div className="dark:bg-black-1 bg-grey-5 mx-2.5 my-3 p-2.5">*/}
             <Container colorScheme={'default'}
-            className={'mx-2.5 my-3 p-2.5 w-auto rounded-[4px]'}
-            >
+                className={'mx-2.5 my-3 p-2.5 w-auto rounded-[4px]'}>
                 <div className="flex justify-between mb-2">
-                    <span className="!font-regular font-semibold dark:text-grey-2 text-grey-1 underline decoration-dotted">Deposit Ratio</span>
+                    <span className="!font-regular font-semibold dark:text-grey-2 
+                        text-grey-1 underline decoration-dotted">
+                        Deposit Ratio
+                    </span>
                     <div className="!font-regular font-semibold dark:text-grey-8 text-black-4 flex flex-row">
                         <Icon src={`img/crypto/${selectedCard?.sourceToken}.svg`} size='sm' className="mr-1" />
                         <span>50% /&nbsp;</span>
@@ -23,16 +23,31 @@ export const ReviewConfirm: FC = (): ReactElement => {
                     </div>
                 </div>
                 <div className="flex justify-between mb-2">
-                    <span className="!font-regular font-semibold dark:text-grey-2 text-grey-1 underline decoration-dotted">Est. 24H Fees</span>
-                    <span className="!font-regular font-semibold dark:text-grey-8 text-black-4">{selectedCard?.volume}</span>
+                    <span className="!font-regular font-semibold underline
+                        dark:text-grey-2 text-grey-1 decoration-dotted">
+                        Est. 24H Fees
+                    </span>
+                    <span className="!font-regular font-semibold dark:text-grey-8 text-black-4">
+                        {selectedCard?.volume}
+                    </span>
                 </div>
                 <div className="flex justify-between mb-2">
-                    <span className="!font-regular font-semibold dark:text-grey-2 text-grey-1 underline decoration-dotted">Pool Fee Rate</span>
-                    <span className="!font-regular font-semibold dark:text-grey-8 text-black-4">{selectedCard?.fees}</span>
+                    <span className="!font-regular font-semibold dark:text-grey-2 
+                        text-grey-1 underline decoration-dotted">
+                        Pool Fee Rate
+                    </span>
+                    <span className="!font-regular font-semibold dark:text-grey-8 text-black-4">
+                        {selectedCard?.fees}
+                    </span>
                 </div>
                 <div className="flex justify-between mb-2">
-                    <span className="!font-regular font-semibold dark:text-grey-2 text-grey-1 underline decoration-dotted">Total Deposit</span>
-                    <span className="!font-regular font-semibold dark:text-grey-8 text-black-4">{selectedCard?.fees}</span>
+                    <span className="!font-regular font-semibold 
+                        dark:text-grey-2 text-grey-1 underline decoration-dotted">
+                        Total Deposit
+                    </span>
+                    <span className="!font-regular font-semibold dark:text-grey-8 text-black-4">
+                        {selectedCard?.fees}
+                    </span>
                 </div>
             </Container>
         </>
