+++ src/pages/FarmV4/StickyFooter.tsx
@@ -17,14 +17,14 @@ type StickyFooterProps = {
   onClaimClick: () => void
 }
 const StickyFooter: FC<StickyFooterProps> = ({
-                                               isDeposit,
-                                               disableActionButton,
-                                               isLoading,
-                                               onActionClick,
-                                               canClaim,
-                                               claimText,
-                                               onClaimClick
-                                             }): ReactElement => {
+  isDeposit,
+  disableActionButton,
+  isLoading,
+  onActionClick,
+  canClaim,
+  claimText,
+  onClaimClick
+}): ReactElement => {
   const { connected } = useWallet()
   const { isMobile } = useBreakPoint()
   const { mode, isDarkMode } = useDarkMode()
@@ -50,8 +50,8 @@ const StickyFooter: FC<StickyFooterProps> = ({
           <Connect fullWidth containerStyle={'h-8.75 z-0'} customButtonStyle={'h-8.75 '} />
         )}
         <Button variant={'outline'}
-                colorScheme={isDarkMode ? 'default' : 'blue'}
-                className={'p-1.5 bg-white'}
+          colorScheme={isDarkMode ? 'default' : 'blue'}
+          className={'p-1.5 bg-white'}
         >
           <Icon src={`img/assets/refresh_${mode}.svg`} size="sm" />
         </Button>
@@ -63,50 +63,50 @@ const StickyFooter: FC<StickyFooterProps> = ({
               className={'bg-white'}
               iconLeft={<Icon src={`img/assets/footer_filter_${mode}.svg`} size="sm" />}
             >
-            <span className="font-bold text-regular text-black-4 dark:text-white">
-              {isNaN(slippage) ? '0.00' : slippage.toFixed(2)}%
-            </span>
+              <span className="font-bold text-regular text-black-4 dark:text-white">
+                {isNaN(slippage) ? '0.00' : slippage.toFixed(2)}%
+              </span>
             </Button>
           </PopoverTrigger>
           <PopoverContent className={`flex flex-col max-sm:w-screen max-sm:rounded-b-none gap-2.5`}
-                          sideOffset={isMobile ?
-                            canClaim ? -90 : -44 :
-                            5}
-                          align={isMobile ? 'center' : 'end'}
-                          alignOffset={0}
+            sideOffset={isMobile ?
+              canClaim ? -90 : -44 :
+              5}
+            align={isMobile ? 'center' : 'end'}
+            alignOffset={0}
           >
             <div className={'flex gap-1 items-center'}>
               <h5>Liquidity Slippage</h5>
               <IconTooltip tooltipType={'outline'}>
-              <span className="font-semibold text-tiny">
-                The maximum slippage that you are willing to accept for this transaction.
-              </span>
+                <span className="font-semibold text-tiny">
+                  The maximum slippage that you are willing to accept for this transaction.
+                </span>
               </IconTooltip>
             </div>
             <RadioOptionGroup defaultValue={'0.1'}
-                              value={localIsCustomSlippage ? 'custom' : value.toString()}
-                              options={[
-                                {
-                                  label: '0.1%',
-                                  value: '0.1',
-                                  onClick: () => setValue(0.1)
-                                },
-                                {
-                                  label: '0.5%',
-                                  value: '0.5',
-                                  onClick: () => setValue(0.5)
-                                },
-                                {
-                                  label: '1%',
-                                  value: '1',
-                                  onClick: () => setValue(1)
-                                },
-                                {
-                                  label: 'Custom',
-                                  value: 'custom',
-                                  onClick: () => setValue(0)
-                                }
-                              ]} />
+              value={localIsCustomSlippage ? 'custom' : value.toString()}
+              options={[
+                {
+                  label: '0.1%',
+                  value: '0.1',
+                  onClick: () => setValue(0.1)
+                },
+                {
+                  label: '0.5%',
+                  value: '0.5',
+                  onClick: () => setValue(0.5)
+                },
+                {
+                  label: '1%',
+                  value: '1',
+                  onClick: () => setValue(1)
+                },
+                {
+                  label: 'Custom',
+                  value: 'custom',
+                  onClick: () => setValue(0)
+                }
+              ]} />
             <Input
               className={'text-right'}
               value={value}
